Fix query mismatch
